Partly done backend for enterSale

+++ DATABASE.sql
@@ -157,7 +157,7 @@ CREATE TABLE purchases (-- Tracks purchases of products from supplier to farm eg
 CREATE TABLE product_inventory (-- Now that products have been produced, how much do we have in the store that hasn't been sold?
     id INT AUTO_INCREMENT PRIMARY KEY,
     product_id INT NOT NULL,
-    quantity_available INT DEFAULT 0,
+    quantity_available DECIMAL(10,2) DEFAULT 0.00,
     last_updated DATETIME DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (product_id) REFERENCES products(id)
 );
